refactor: for new model
