+++ dataprofiler/labelers/char_load_tf_model.py
@@ -19,254 +19,22 @@ from .. import dp_logging
 _file_dir = os.path.dirname(os.path.abspath(__file__))
 
 logger = dp_logging.get_child_logger(__name__)
+labeler_utils.hide_tf_logger_warnings()
 
 
-class NoV1ResourceMessageFilter(logging.Filter):
-    """Removes TF2 warning for using TF1 model which has resources."""
-    def filter(self, record):
-        msg = 'is a problem, consider rebuilding the SavedModel after ' + \
-            'running tf.compat.v1.enable_resource_variables()'
-        return msg not in record.getMessage()
-
-
-tf_logger = logging.getLogger('tensorflow')
-tf_logger.addFilter(NoV1ResourceMessageFilter())
-
-
-def protected_register_keras_serializable(package='Custom', name=None):
-    """
-    Protects against already registered keras serializable layers. This
-    ensures that if it was already registered, it will not try to register it
-    again.
-    """
-    def decorator(arg):
-        """Protects against double registration of a keras layer."""
-        class_name = name if name is not None else arg.__name__
-        registered_name = package + '>' + class_name
-        if tf.keras.utils.get_registered_object(registered_name) is None:
-            tf.keras.utils.register_keras_serializable(package, name)(arg)
-        return arg
-    return decorator
-
-
-@protected_register_keras_serializable()
-class FBetaScore(tf.keras.metrics.Metric):
-    r"""Computes F-Beta score.
-    Adapted and slightly modified from https://github.com/tensorflow/addons/blob/v0.12.0/tensorflow_addons/metrics/f_scores.py#L211-L283
-
-    # Copyright 2019 The TensorFlow Authors. All Rights Reserved.
-    #
-    # Licensed under the Apache License, Version 2.0 (the "License");
-    # you may not use this file except in compliance with the License.
-    # You may obtain a copy of the License at
-    #
-    #     http://www.apache.org/licenses/LICENSE-2.0
-    #     https://github.com/tensorflow/addons/blob/v0.12.0/LICENSE
-    #
-    # Unless required by applicable law or agreed to in writing, software
-    # distributed under the License is distributed on an "AS IS" BASIS,
-    # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    # See the License for the specific language governing permissions and
-    # limitations under the License.
-    # ==============================================================================
-
-    It is the weighted harmonic mean of precision
-    and recall. Output range is `[0, 1]`. Works for
-    both multi-class and multi-label classification.
-    $$
-    F_{\beta} = (1 + \beta^2) * \frac{\textrm{precision} * \textrm{precision}}{(\beta^2 \cdot \textrm{precision}) + \textrm{recall}}
-    $$
-    Args:
-        num_classes: Number of unique classes in the dataset.
-        average: Type of averaging to be performed on data.
-            Acceptable values are `None`, `micro`, `macro` and
-            `weighted`. Default value is None.
-        beta: Determines the weight of precision and recall
-            in harmonic mean. Determines the weight given to the
-            precision and recall. Default value is 1.
-        threshold: Elements of `y_pred` greater than threshold are
-            converted to be 1, and the rest 0. If threshold is
-            None, the argmax is converted to 1, and the rest 0.
-        name: (Optional) String name of the metric instance.
-        dtype: (Optional) Data type of the metric result.
-    Returns:
-        F-Beta Score: float.
-    """
-
-    # Modification: remove the run-time type checking for functions
-    def __init__(self, num_classes, average=None, beta=1.0, threshold=None,
-                 name="fbeta_score", dtype=None, **kwargs):
-        super().__init__(name=name, dtype=dtype)
-
-        if average not in (None, "micro", "macro", "weighted"):
-            raise ValueError(
-                "Unknown average type. Acceptable values "
-                "are: [None, 'micro', 'macro', 'weighted']"
-            )
-
-        if not isinstance(beta, float):
-            raise TypeError("The value of beta should be a python float")
-
-        if beta <= 0.0:
-            raise ValueError("beta value should be greater than zero")
-
-        if threshold is not None:
-            if not isinstance(threshold, float):
-                raise TypeError("The value of threshold should be a python float")
-            if threshold > 1.0 or threshold <= 0.0:
-                raise ValueError("threshold should be between 0 and 1")
-
-        self.num_classes = num_classes
-        self.average = average
-        self.beta = beta
-        self.threshold = threshold
-        self.axis = None
-        self.init_shape = []
-
-        if self.average != "micro":
-            self.axis = 0
-            self.init_shape = [self.num_classes]
-
-        def _zero_wt_init(name):
-            return self.add_weight(
-                name, shape=self.init_shape, initializer="zeros", dtype=self.dtype
-            )
-
-        self.true_positives = _zero_wt_init("true_positives")
-        self.false_positives = _zero_wt_init("false_positives")
-        self.false_negatives = _zero_wt_init("false_negatives")
-        self.weights_intermediate = _zero_wt_init("weights_intermediate")
-
-    def update_state(self, y_true, y_pred, sample_weight=None):
-        if self.threshold is None:
-            threshold = tf.reduce_max(y_pred, axis=-1, keepdims=True)
-            # make sure [0, 0, 0] doesn't become [1, 1, 1]
-            # Use abs(x) > eps, instead of x != 0 to check for zero
-            y_pred = tf.logical_and(y_pred >= threshold, tf.abs(y_pred) > 1e-12)
-        else:
-            y_pred = y_pred > self.threshold
-
-        y_true = tf.cast(y_true, self.dtype)
-        y_pred = tf.cast(y_pred, self.dtype)
-
-        def _weighted_sum(val, sample_weight):
-            if sample_weight is not None:
-                val = tf.math.multiply(val, tf.expand_dims(sample_weight, 1))
-            return tf.reduce_sum(val, axis=self.axis)
-
-        self.true_positives.assign_add(_weighted_sum(y_pred * y_true, sample_weight))
-        self.false_positives.assign_add(
-            _weighted_sum(y_pred * (1 - y_true), sample_weight)
-        )
-        self.false_negatives.assign_add(
-            _weighted_sum((1 - y_pred) * y_true, sample_weight)
-        )
-        self.weights_intermediate.assign_add(_weighted_sum(y_true, sample_weight))
-
-    def result(self):
-        precision = tf.math.divide_no_nan(
-            self.true_positives, self.true_positives + self.false_positives
-        )
-        recall = tf.math.divide_no_nan(
-            self.true_positives, self.true_positives + self.false_negatives
-        )
-
-        mul_value = precision * recall
-        add_value = (tf.math.square(self.beta) * precision) + recall
-        mean = tf.math.divide_no_nan(mul_value, add_value)
-        f1_score = mean * (1 + tf.math.square(self.beta))
-
-        if self.average == "weighted":
-            weights = tf.math.divide_no_nan(
-                self.weights_intermediate, tf.reduce_sum(self.weights_intermediate)
-            )
-            f1_score = tf.reduce_sum(f1_score * weights)
-
-        elif self.average is not None:  # [micro, macro]
-            f1_score = tf.reduce_mean(f1_score)
-
-        return f1_score
-
-    def get_config(self):
-        """Returns the serializable config of the metric."""
-
-        config = {
-            "num_classes": self.num_classes,
-            "average": self.average,
-            "beta": self.beta,
-            "threshold": self.threshold,
-        }
-
-        base_config = super().get_config()
-        return {**base_config, **config}
-
-    def reset_states(self):
-        reset_value = tf.zeros(self.init_shape, dtype=self.dtype)
-        tf.keras.backend.batch_set_value([(v, reset_value) for v in self.variables])
-
-
-@protected_register_keras_serializable()
-class F1Score(FBetaScore):
-    r"""Computes F-1 Score.
-
-    # Copyright 2019 The TensorFlow Authors. All Rights Reserved.
-    #
-    # Licensed under the Apache License, Version 2.0 (the "License");
-    # you may not use this file except in compliance with the License.
-    # You may obtain a copy of the License at
-    #
-    #     http://www.apache.org/licenses/LICENSE-2.0
-    #     https://github.com/tensorflow/addons/blob/v0.12.0/LICENSE
-    #
-    # Unless required by applicable law or agreed to in writing, software
-    # distributed under the License is distributed on an "AS IS" BASIS,
-    # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    # See the License for the specific language governing permissions and
-    # limitations under the License.
-    # ==============================================================================
-
-    It is the harmonic mean of precision and recall.
-    Output range is `[0, 1]`. Works for both multi-class
-    and multi-label classification.
-    $$
-    F_1 = 2 \cdot \frac{\textrm{precision} \cdot \textrm{recall}}{\textrm{precision} + \textrm{recall}}
-    $$
-    Args:
-        num_classes: Number of unique classes in the dataset.
-        average: Type of averaging to be performed on data.
-            Acceptable values are `None`, `micro`, `macro`
-            and `weighted`. Default value is None.
-        threshold: Elements of `y_pred` above threshold are
-            considered to be 1, and the rest 0. If threshold is
-            None, the argmax is converted to 1, and the rest 0.
-        name: (Optional) String name of the metric instance.
-        dtype: (Optional) Data type of the metric result.
-    Returns:
-        F-1 Score: float.
-    """
-
-    # Modification: remove the run-time type checking for functions
-    def __init__(self, num_classes, average=None, threshold=None,
-                 name="f1_score", dtype=None):
-        super().__init__(num_classes, average, 1.0, threshold, name=name, dtype=dtype)
-
-    def get_config(self):
-        base_config = super().get_config()
-        del base_config["beta"]
-        return base_config
-
-
-class CharLoadTFCnnModel(BaseTrainableModel,
-                         metaclass=AutoSubRegistrationMeta):
+class CharLoadTFModel(BaseTrainableModel,
+                      metaclass=AutoSubRegistrationMeta):
 
     # boolean if the label mapping requires the mapping for index 0 reserved
     requires_zero_mapping = False
 
-    def __init__(self, original_model_path, label_mapping=None,
+    def __init__(self, model_path, label_mapping=None,
                  parameters=None):
         """
-        CNN Model Initializer. initialize epoch_id
+        Loadable TF Model Initializer.
 
+        :param model_path: path to model to load
+        :type model_path: str
         :param label_mapping: maps labels to their encoded integers
         :type label_mapping: dict
         :param parameters: Contains all the appropriate parameters for the
@@ -281,7 +49,7 @@ class CharLoadTFCnnModel(BaseTrainableModel,
         if not parameters:
             parameters = {}
         parameters.setdefault('default_label', "UNKNOWN")
-        parameters['original_model_path'] = original_model_path
+        parameters['model_path'] = model_path
         parameters['pad_label'] = 'PAD'
         self._epoch_id = 0
 
@@ -327,12 +95,12 @@ class CharLoadTFCnnModel(BaseTrainableModel,
         :return: None
         """
         errors = []
-        list_of_necessary_params = ['original_model_path', 'default_label',
+        list_of_necessary_params = ['model_path', 'default_label',
                                     'pad_label']
 
         # Make sure the necessary parameters are present and valid.
         for param in parameters:
-            if param in ['default_label', 'original_model_path', 'pad_label']:
+            if param in ['default_label', 'model_path', 'pad_label']:
                 if not isinstance(parameters[param], str):
                     error = str(param) + " must be a string."
                     errors.append(error)
@@ -426,7 +194,7 @@ class CharLoadTFCnnModel(BaseTrainableModel,
 
         # use f1 score metric
         custom_objects = {
-            "F1Score": F1Score(
+            "F1Score": labeler_utils.F1Score(
                 num_classes=max(label_mapping.values()) + 1,
                 average='micro'),
             "CharacterLevelCnnModel": cls,
@@ -444,11 +212,6 @@ class CharLoadTFCnnModel(BaseTrainableModel,
         ]
         return loaded_model
 
-    def _get_layer_index(self, layer_name):
-        for idx, layer in enumerate(self._model.layers):
-            if layer.name == layer_name:
-                return idx
-
     def _construct_model(self):
         """
         Model constructor for the data labeler. This also serves as a weight
@@ -458,11 +221,12 @@ class CharLoadTFCnnModel(BaseTrainableModel,
         """
         num_labels = self.num_labels
         default_ind = self.label_mapping[self._parameters['default_label']]
-        model_loc = self._parameters['original_model_path']
+        model_loc = self._parameters['model_path']
 
         self._model = tf.keras.models.load_model(model_loc)
         softmax_output_layer_name = self._model.outputs[0].name.split('/')[0]
-        softmax_layer_ind = self._get_layer_index(softmax_output_layer_name)
+        softmax_layer_ind = labeler_utils.get_tf_layer_index_from_name(
+            self._model, softmax_output_layer_name)
         softmax_layer = self._model.get_layer(softmax_output_layer_name)
         prev_softmax_layer = softmax_layer.input
 
@@ -484,7 +248,8 @@ class CharLoadTFCnnModel(BaseTrainableModel,
         losses = {softmax_output_layer_name: "categorical_crossentropy"}
 
         # use f1 score metric
-        f1_score_training = F1Score(num_classes=num_labels, average='micro')
+        f1_score_training = labeler_utils.F1Score(
+            num_classes=num_labels, average='micro')
         metrics = {softmax_output_layer_name: ['acc', f1_score_training]}
 
         self._model.compile(loss=losses, optimizer="adam", metrics=metrics)
@@ -536,7 +301,8 @@ class CharLoadTFCnnModel(BaseTrainableModel,
         losses = {softmax_output_layer_name: "categorical_crossentropy"}
 
         # use f1 score metric
-        f1_score_training = F1Score(num_classes=num_labels, average='micro')
+        f1_score_training = labeler_utils.F1Score(
+            num_classes=num_labels, average='micro')
         metrics = {softmax_output_layer_name: ['acc', f1_score_training]}
 
         self._model.compile(loss=losses, optimizer="adam", metrics=metrics)
